Address code review suggestion of doing glob and copy asynchronously
* Use async glob and fsCopy to find and fill in missing frames, integrating into the
existing videoPromise

* amend test and test coverage

* attach the video placeholder to allure earlier (see below)

* Amend helpers.sleep() test to apply to Atomics.wait()

* Fix rejected screenshot promise handling

Turns out doing it this way revealed a lot more fragilities in the workflow. It
looks like the missing frame code was not being triggered properly because
browser.screenshot() was not being treated asynchronously.

Also allure.addAttachment() would sometimes be called after the start of the
next test causing the video to be assigned to the wrong test or to be missing in
Allure altogether, despite being successfully recorded, so a different strategy
has been implemented to add the attachment as early as possible - before the first
command of the test.